fix: propose restructure
