+++ test/OptimisticDepositBox/OptimisticDepositBox.Withdraw.ts
@@ -1,6 +1,6 @@
-import { SignerWithAddress, expect, Contract, ethers } from "./utils";
-import { umaEcosystemFixture } from "./fixtures/UmaEcosystem.Fixture";
-import { optimisticDepositBoxFixture } from "./fixtures/OptimisticDepositBox.Fixture";
+import { SignerWithAddress, expect, Contract, ethers } from "../utils";
+import { umaEcosystemFixture } from "../fixtures/UmaEcosystem.Fixture";
+import { optimisticDepositBoxFixture } from "../fixtures/OptimisticDepositBox.Fixture";
 import {
   amountToSeedWallets,
   amountToDeposit,
@@ -9,12 +9,12 @@ import {
   zeroBytes,
   mockPrice,
   identifier,
-} from "./constants";
+} from "../constants";
 
 let optimisticDepositBox: Contract, usdc: Contract, timer: Contract, optimisticOracle: Contract;
 let deployer: SignerWithAddress, depositor: SignerWithAddress, proposer: SignerWithAddress;
 
-describe("Optimistic Deposit Box Withdraw functions", function () {
+describe("Optimistic Deposit Box: Withdraw", function () {
   beforeEach(async function () {
     [deployer, depositor, proposer] = await ethers.getSigners();
     ({ timer, optimisticOracle } = await umaEcosystemFixture());
