+++ test/fixtures/index.ts
@@ -1,0 +1,3 @@
+export * from "./EventBasedPredictionMarket.Fixture";
+export * from "./OptimisticDepositBox.Fixture";
+export * from "./UmaEcosystem.Fixture";
